Working Torch_BP added
When hit (only) with a fireball, it'll activate similar to a button. I tried it with the door camera angle, but it would glitch out so I just removed it for the torches.

I'm sure in the future there will be instances where multiple torches will need to be lit, but that's a story for another day.